+++ controllers/topicsController.js
@@ -1,0 +1,12 @@
+const {getAllTopics} = require('../models/topicModel')
+exports.getTopics =(req,res,next)=>{
+    getAllTopics()
+.then((topics)=>{
+        
+       res.status(200).send(topics) 
+})
+.catch(err => {
+    next(err);   
+})
+
+}
