+++ endpoints.json
@@ -82,6 +82,20 @@
     "votes": 0,
     "created_at": "2024-04-18T15:51:12.835Z"
   }
+},
+"POST /api/articles/:article_id": {
+  "description": "Updated votes according the article_id with inc_votes send to the server ",
+  "queries": ["article_id","inc_votes"],
+  "exampleResponse": {
+    "article_id": 1,
+        "title": "Living in the shadow of a great man",
+        "topic": "mitch",
+       " author": "butter_bridge",
+        "body": "I find this existence challenging",
+      "  created_at": "2020-07-09T20:11:00.000Z",
+        "votes": 110,
+       " article_img_url": "https://images.pexels.com/photos/158651/news-newsletter-newspaper-information-158651.jpeg?w=700&h=700"
+  }
 }
 }
